HTTP_service/requirements.txt
Django==4.2.5
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
pandas==2.1.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1





